App: move plugins
